Make assignHandlers a generator
Our intent with assignHandlers is to run it almost continuously,
but in case it is slow, to occasionally remove completed handlers
and run paused handlers in order to make room for new requests
which may otherwise be starved.

The way that currently works is that we stop processing requests
after 15 seconds, allow the other methods to run, then start over.
If this happens continuously, we may never see requests near the
bottom of the list (which we might be able to satisfy).  To avoid
that, this change turns assignHandlers into a generator which
picks up where it left off each time we yield to the other
processors.

Change-Id: I32096ae7342cc8aafd2a14de79acc4267293349a