Fix race in dib_image_list_json test
Change Ibff0a9016936b461eccb1b48dcf42f5ad8d8434e had an error that
was not caught by testing due to a race in the test code -- the
build might finish before the webapp starts.  This corrects the
test (and a few sibling tests) by starting the webapp first, then
starting the builder.

The actual error is also corrected: ImageBuild.updateFromDict is
implemented.  Further, duplicate fromDict method calls are removed
from several classes and renamed to updateFromDict so that the
APIs match.

Change-Id: I918d1badf838a23d6b0813b5b31dac4888d04ce0