Use node cache in node deleter
The node deleter method runs every 5 seconds and iterates over
every node.  For any node that is not READY, it attempts to lock
the node to determine if it should be deleted.  In other words,
this method attempts to lock almost every node every 5 seconds.

Locking nodes is expensive (involving a number of ZK round trips).
However, our node cache keeps a count of the number of lock
contenders on a given node.  That can be used as a proxy for whether
the node is probably locked.  Zero contenders means it could be
unlocked.  One or more contenders means it's probably locked.

To greatly reduce ZK traffic, we can use the node cache and only
attempt to lock nodes if the cache indicates they are potentially
unlocked.  If we're wrong, we'll try again in 5 seconds anyway.

Change-Id: Ieb54babef92d5dbf3173316c6a1711e0e4a70403