Add a LazyExecutorTTLCache to the OpenStack driver
See the docstring for an explanation of what a Lazy Executor TTL
Cache is.  By switching the caching of the server list method (and
also volumes and fips) to the lazy cache, we will make all of the
methods called by the state machines asynchronous.  This means
that both the create and delete state machine threads should be
able to spin through all of their state machines as quickly as
Python and ZooKeeper overhead will allow.

Change-Id: Ibce6b4d82929e6a764fdbc025990f7e01060b509