Mutex access to local lock attributes
When we lock or unlock requests or nodes, we set the .lock attribute
of the request/node to the ZK lock object or None respectively.

If two threads are racing an unlock and lock, we can have the following
sequence:

1: unlock request in ZK
2: lock request in ZK
2: set request.lock to ZK lock object
1: set request.lock object to None

If this happens, thread 2 will not be able to unlock the request
and it will stay locked permanently.

To correct this, set a local mutex around these operations so that
they can not be sequenced as above.

Change-Id: Ic52dc0202bf629462d774406b08b7dee01396dd8