Avoid python/zk lock deadlocks
Change Ic52dc0202bf629462d774406b08b7dee01396dd8 added a thread lock
to the Node and NodeRequest objects to avoid multiple threads
overwriting the lock attribute and therefore causing us to forget
whether a node was locked.  However, some lock paths employ blocking
zk locks which can lead to deadlocks between the two types of locks.

I.e., one thread holds the ZK lock and tries to acquire the thread
lock in order to release the ZK lock, while another thread holds
the thread lock while trying to acquire the ZK lock.

We could avoid all of this if we get rid of the two-phase commit
problem where we set the lock attribute of the Node after
acquiring or releasing the ZK lock and just rely on the ZK lock
itself.  However, the underlying kazoo Lock recipe appears to
suffer from the same original issue: it appears possible for an
release to overwrite the internal locked flag while an acquire is
happening.

So instead, we shore up our locking strategy by holding the
thread lock over the entire time period that the ZK lock is held.
This means that locking order is always:

1) acquire thread
2) acquire zk
3) release zk
4) release thread

Except in the case of a non-ephemeral ZK lock, such as that used
by the metastatic driver.  We release the thread lock after
acquiring the ZK lock in that case.  It doesn't really matter
anyway, the lock is never used again, and the ZK unlock method is
performed without reference to the Node.lock attribute.  But it
seems like the right behavior.

Change-Id: I01b14a5af83ab6067573ef1dee9803041c339fbe