Reduce sleeps to improve time to ready
The statemachine driver currently sleeps 10 seconds between each
iteration of the loop that runs all the state machines.  This time
can delay the time to ready for nodes if the cloud is relatively
fast.  This change reduces the delay to a max of 1 second (if the
loop takes longer than 1 second, the delay will be 0).

Additionally, the Lazy TTL cache used in the openstack driver has
a TTL of 10 seconds.  This can contribute to long times to ready
as well.  This change reduces the TTL to 1 second (we will wait
at least 1 second after the completion of a list servers call before
we fire off the next one).  Because list servers can take quite some
time, this could cause one of the threadpoolexecutor threads to spend
more time waiting for a list servers call to return.  However, local
load testing with 600 outstanding requests shows no impact to
overall throughput.

Finally, in a semi-related change, the launcher watermark sleep
(the time that the launcher waits between loop iterations to check
for new handlers and completed handlers) is reduced as well.  Now
that most of the launcher work operates on local tree cache data,
we can run this more often without incurring additional ZK load.

Given a hypothetical cloud that responds instantaneously, we should
be able to satisfy a request in about 3 seconds.

The first two changes improve response time without lowering
throughput, the third change improves both.

These will output considerably more debug log messages.

Change-Id: Idaad3de2a2afb5d51ede680ecf33c1d5c62fbdbb