Release the thread lock when we delete a node
Node locks in ZK are held underneath the node znode, which means they
are automatically "released" (i.e., deleted) when the node is deleted.
Because of that, when we call deleteNode, we don't call unlockNode
subsequently.

But that fails to release the local python _thread_lock, so if another
thread uses the cached Node object, it will try to acquire that lock
and deadlock.

This updates deleteNode to release the local _thread_lock if necessary.

Change-Id: I3424b1ef54ca6678da05c7d861bd96be5343c5a4