Add Azure gallery image support
The shared and community gallery images are another way to specify
what image to use when creating a VM.  Shared galleries are intended
for use within an organization, and community galleries are public.

This adds support for using these images.  It requires an API version
bump since the virtual machine attributes to specify them are new.

Change-Id: Ia981fcbeea6680a9d14ee8e4ec401bf227a7cc12