Add missing request unlock
Because we now iterate over cached requests, when we decide that
we should lock a request to attempt to handle or decline it, we
double check the status of the request after locking it.  The act
of locking the request refreshes the data from ZK, so at this point
we are certain that the data are up to date.

In the case where, after locking and refreshing, we double checked
that we had not already declined the request, we are missing an
instruction to unlock the request.  This can leave a request
perpetually locked.

This change adds the missing unlock.

Change-Id: Ifcf52bbca03329a8ca7015412f9aaf795c5ae7c0