Fix connection handling in recursive tree cache
There were two errors related to connection handling in the new
treecache implementation:

1) Upon disconnection, the cache receives a "NONE" event with
the client state ("CONNECTING") and no path.  This is a signal to
the watcher that the connection state has changed.  Since we process
connection changes via the session listener, we should ignore this
event.  We were previously attempting to process it and assumed that
the path attribute would be present.  This change ignores NONE events
with no path.

2) The treecache is responsible for re-establishing watches if the
connection is LOST or SUSPENDED.  Previously, it would only
re-establish watches and refresh the cache if the connection was LOST.
This change updates it to re-establish/refresh in both cases.  The
"_started" variable is no longer necessary to distinguish between the
two.  However, it may be useful for future consumers to know whether
or not the cache is synced, so it is converted to "_ready" in case
we want to expose that in the future.

Change-Id: Id1f9c34628d1cc04881621a83d1d1f78e9e0f366