Refactor statsd client usage to consolidate presence checking
Everywhere we use the statsd client we need to check if it is None or a
valid object. This is due to statsd being an optional configuration item
for nodepool. Encapsulate the actual statsd client in order to move the
checking into a more centralized location. This should help us avoid
bugs where we forget tocheck the client is valid before using it. Also
reduces block nesting.

Change-Id: I97665718c5792413725f3d4f2bdab4d7b8e29c0b