Improve tree cache resiliency at startup
This makes the following changes:

1) It wraps the cache initialization in a lock so that in the
unlikely event there is a zookeeper disconnect between the time
that the connection listener is registered and the initial sync
is commanded, we won't end up with two threads repopulating the
cache at the same time.

2) It ensures that errors in the _start method don't cause a
traceback that would scuttle the construction of the cache.  This
is because we construct 3 cache objects in a row on the ZooKeeper
object, and if there is an error during startup, The nodepool
launcher will be fine but we could end up leaking entire caches
because the kazoo client has a session listener reference[1].
This change ensures that we get all the way through _start even
if there's an error.

3) It halts use of the cache if the cache is not ready.  All
references to the cache now use API methods that will wait until
the cache is ready before returning any data.  This means that
if there is a connection error during startup or any other time,
we will block until the cache is back in sync.  Nodepool has its
own methods for stopping processing during a connection outage,
but this additional step ensures we won't perform processing with
incomplete data (e.g., a list of images that is too short).

4) Some improved logging.

[1] kazoo should probably use weakrefs when maintining its connection
listeners, but it does not.  We could probably fix that ourselves
with a subclass.

Change-Id: I4045b332ca14bc974d67a78866bad304aabac590