Retry launches in statemachine driver
This adds retry handling to the state machine driver and removes it
from the driver adapters.

This allows the adapters to be simpler as they don't need to implement
any internal retry handling.  It also lets us choose to retry due
to events handled by the statemachine driver itself.

One of those events is a failed keyscan.  In the old OpenStack driver,
a failed keyscan would result in a retry rather than a launch failure,
but the migration of the OpenStack driver to statemachine inadvertently
changed that behavior.  This restores the old behavior, applies it
to all drivers, and codifies it with a unit test.

Several logging lines have been changed to use arguments instead of
format strings for consistency and safety.

Change-Id: I670ef8d762558cf8346f071be282af8dfc803747