Defer node request when label is not available
During assignHandlers we already gather the pools that can serve a
label. In case the pool doesn't offer a label of the request the
current process is that we lock and decline the node request right
away. In a system where there are many pools that don't satisfy a
label most requests typically get locked and declined by many pools
before they get served by a pool that supports a label.

This changes the behavior in a way that request handling is deferred
if the label is not supported on the pool but we know that there is
another pool that supports it. This way the pools that are in charge
for the label will be the first ones to lock and process the
request. If there is no candidate launcher pool anymore the request
will still be processed and declined in the nextassign handlers loop.

In summary this should reduce load on zookeeper and at the same time
improve performance of the assignHandlers loop by avoiding locking in
potentially a lot of cases.

Change-Id: Ic48b87218470d3c652c1e0b3815298354789efea