Move cache event logs to their own logger
On high traffic systems, the cache event logs can be very high
volume and also not very useful.  Similar to how we put zuul
connection io traffic in its own logger so that folks can filter
it out, let's move the cache event log messages to their own.

Eventually we should probably remove this log line entirely, but
since the cache is new, we may yet need to debug basic functionality,
so putting it in its own logger lets us turn it on and off easily.

Change-Id: Ice382748a0b21a7e73bb0d2b3461761a8d7871da