Update patch set 2
Patch Set 2:

To be 100% clear on this: running multiple launchers against the same provider is not supported by nodepool.

However, this is a simple change with no performance impact or increased complexity, and it's a straightforward additional safety check that is good regardless.  Additionally, we would, in the future, like to be able to support that use case, so additional prep work that doesn't overly complicate the existing implementation should be welcome.  So I think this is a good change to merge.  I just want to be clear that merging this doesn't mean that use case is supported yet.  Anyone doing that is blazing their own trail.

Patch-set: 2
CC: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>