Add timing information to "Node is ready" message
So far we have statsd metrics that might show startup times of
nodes. However it can be hard to find those in the logs because the
timing is not part of the "Node is ready" message so we have to grep
the logs for many individidual nodes and calculate the timings.

This change makes it easier to spot nodes that were slow so we can
analyse them easier on openstack level.

Change-Id: Ic42b8893148a47d5dce84f9739ed38fae2d6b207