Parallelize static startup more
We intend to parallelize the startup of the static driver because
it involves a number of network checks which may slow down startup
if there are a lot of hosts.  However, the current implementation
only does that for half of the startup checks.

When we startup we:

1) Register any nodes defined in the config that don't already
   exist in ZK.
2) Update any nodes that do exist in ZK with any changes from the
   config.

Notably, in all cases at startup we perform steps 1 and 2, so new
nodes get written to ZK twice.  Step 1 is parallelize in an executor,
step 2 is not.

This change makes an update so that both steps are parallelized,
and step 2 is run only if step 1 is not.

Step 1 corresponds to syncNodeCount() and step 2 is
updateNodeFromConfig().

Note that they can't be combined directly because syncNodeCount is
called in other code paths (the delete/reregister path) but
updateNodeFromConfig is only needed at startup (keep in mind that
an online config change is seen as a new startup -- the provider
object is idled and a new provider started).

Change-Id: I80822757a9378138919c9f4157641b4b76014db2