Use cache more aggressively when searching for ready nodes
When assigning node requests, the pool main loop attempts to find
existing ready nodes for each request it accepts before creating
a NodeLauncher (which is the class which usually causes asynchronous
node creation to happen).

Even though we iterate over every ready node in the system looking
for a match, this should be fast because:

* We only consider nodes from our provider+pool.
* We only consider nodes which are not already assigned (and newly
  created nodes are created with assigned_to already set).
* We use the node cache.

However, there is room for improvement:

* We could use cached_ids so that we don't have to list children.
* We can avoid trying to lock nodes that the cache says are already
  locked (this should generally be covered by the "assigned_to" check
  mentioned above, but this is a good extra check and is consistent
  with other new optimizations.
* We can ignore nodes in the cache with incomplete data.

The first and last items are likely to make the most difference.

It has been observed that in a production system we can end up spending
much more time than expected looking for ready nodes.  The most likely
cause for this is cache misses, since if we are able to use cached data
we would expect all the other checks to be dealt with quickly.  That
leaves nodes that have appeared very recently (so they show up in a
live get_children call but aren't in the cache yet, or if they are,
are in the cache with incomplete (empty) data) or nodes that have been
disappeared very recently (so they showed up in our initial listing
but by the time we get around to trying to lock them, they have been
removed from the cache).

This also adds a log entry to the very chatty "cache event" log handler
in the case of a cache miss in order to aid future debugging.

Change-Id: I32a0f2d0c2f1b8bbbeae1441ee48c8320e3b9129