Use asynchronous fetch operations for tree cache
The TreeCache can be a bit of a bottleneck since it is centralizing
in a single thread fetch operations which may have previously
happened in different threads.  It also may in some cases fetch
more data than we did previously since it will update all objects
in memory as they are updated in ZK regardless of whether nodepool
is interested in them at the moment.  Or at least, it may concentrate
those fetches in a smaller timespan.

To address the bottleneck, this change uses asynchronous fetches
from ZooKeeper.  The background worker which previously updated the
cache in response to events from ZK is split into two threads.  The
first will decide whether or not a fetch for a given event is
necessary, and then fire off the async fetch operation if so.  It
then pushes the future for that operation, along with other event
information into a second queue.  A second background thread
(called the "playback" worker) will pop events from that queue,
still in order, and await the future.  This ensures all cache
operations are still ordered, while multiple fetch operations can
be outstanding in parallel.

This relies even more heavily on the path parsing implemented by the
three caches to be correct.  In particular, we want to make sure that
we don't accidentally fetch a znode if we're just dealing with a
lock node event.  The previous code sequenced the checks so that we
would handle the lock node event and short circuit before deciding
whether to fetch.  The new code reverses that, which means that it
is paramount that we don't accidentally have the path parsing return
a match on a lock node.  The cache would function correctly, but it
would silently perform far more operations than necessary, so such an
error would likely go unnoticed for some time.  To prevent this,
explicit testing for path parsing is added.

Change-Id: I6ef10c724c3993ee565510ab917dce64d2e3d3f9