Correct node cache lock handling
A logic error in the TreeCache prevented us from correctly
re-establishing the state of node locks after a zk connection
loss.  In such a case, we need to fetch data for every item in
the tree to determine whether it exists, but we were only fetching
data for cache objects, not ancillary znodes like lock nodes.

Change-Id: I4e9faa4981b90a6b1f46d2f0347866e2166d5b7f