Clear tree cache queues on disconnect
If a TreeCache encounters a ZK disconnect, then any events in its
event queue are no longer useful and possibly counter-productive.
If we process events from before the disconnection after a
reconnection, the cache could get into an erroneous state.

The cache is designed so that on initial or re-connection, we first
establish the watch, and then walk the tree to look for previously
existing data.  All events we process (whether generated by the
watch or by walking the tree) must be in order after the watch was
established.

To ensure this, we clear the event queues on reconnect and restart
the event queue threads.

Additionally, this change cleans up handling of the case where
the cache key is None (the parser says we're not interested in the
object) but the preCacheHook didn't tell us to exit early.  In that
case, we would previously most likely try to pop the None object
from the cache, which would generally be harmless.  But we shouldn't
be doing that anyway, so add a check for whether we have an object
key in the main cache method and ignore the return values from the
preCacheHook.

Change-Id: I4bbedb19364c894a2033ef8c1d2e439299971a83