Log and re-raise cache startup errors
We've seen cases where the cache failed to start but no related
exception could be found in the logs. This is now changed to also
include the traceback

In cases where the cache init fails, the Nodepool launcher will just
hang there, waiting for the cache to be ready. To avoid such a situation
we'll re-raise the exception.

Change-Id: I3bbc83ec8c3fbb382e3e39a84f6f5bad7dc2ee41