Limit TreeCache playback queue size
The playback queue includes (but is not limited to) the outstanding
requests to ZooKeeper.  If we allow it to grow too large, we may
overwhelm the zk client connection (it uses a unix socket for
intra-thread communication, and if we have more outstanding requests
than the socket buffer size, we may block on submitting new requests
to the client).  This should also prevent the cache from monopolizing
the client connection.

Once the playback size is limited, that will cause the event queue
worker to block when moving items from the event queue to the playback
queue.  That will cause the event queue to become larger (essentially
containing almost the entire backlog of events instead of it being
split between the event and playback queues).

That makes the event queue a good place to perform deduplication.
The very common pattern of:

* create node record
* update node record
* update node record
* update node record
* delete node record

can easily be collapsed into as few as 2 events (create / delete)
instead of 5.

Change-Id: I081b32291f53ba18a31f8137749bb5fcb7f84606