Static driver: Update nodes in parallel
When starting the static provider, it concurrently syncs the static
nodes in order to parallelize potentially costly host key scans.
Right after the initial sync, we update known nodes to track config
updates. Host key scanning is done there as well, but not concurrently.
This change makes sure we also do this second round of host checking in
parallel.

Change-Id: I7522cb60a8af56de0369fd16a839465dcabff61c