Don't hold allocated nodes
The hold command currently only checks whether it can lock a node,
but it does not check whether the node is allocated.  The nodepool
protocol requires that we not manipulate nodes unless they are
locked and unallocated or allocated to us.

This means that a hold command could set an allocated node to HOLD
which could then be used as normal by the requestor, effectively
ignoring and overwriting the hold, contrary to user expectation.

To correct this, we check for allocation before and after locking
the node and refuse to hold the node in that situation.

Change-Id: Ie42ec4a33f28a02320855df09fa898200d6ca44e