Don't reload config when file wasn't modified
The launcher thread will reload the config in the main loop every second
(watermark sleep) without checking if the config changed. This can take
quite a bit of CPU time as we've seen from different Yappi profiles.

This change uses the last modified timestamp of the config file to
detect if the Nodepool config needs to be reloaded.

Change-Id: I93400cc156d09ea1add4fc753846df923242c0e6