Increase state machine interval and cache ttl
This increases the state machine poll interval to 2.5 seconds, and
the openstack ttl cache to 5s.  This reduces the amount of cpu
and network traffic on the launcher at the expense of increasing
the minimum time to ready from 3 seconds (on an infinitely fast
cloud) to 8.5 seconds.

(poll interval + cache interval + watermark sleep time=1)

The poll interval is set to half of the ttl under the assumption
that the request in a cache update will usually take a couple of
seconds, so we may be able to increase the amount of work the
launcher can do between cache updates by checking for updated data
more often than the data actually updates.

Change-Id: Ida96142cb04c0dff2d7beb3cd1c3c52a0ee42f5a