Handle invalid AWS instance in quota lookup
Early nodepool performed all cloud operations within the context of
an accepted node request.  This means that any disagreement between
the nodepool configuration and the cloud (such as what instance types,
images, networks, or other resources are actually available) would
be detected within that context and the request would be marked as
completed and failed.

When we added tenant quota support, we also added the possibility
of needing to interact with the cloud before accepting a request.
Specifically, we may now ask the cloud what resources are needed
for a given instance type (and volume, etc) before deciding whether
to accept a request.  If we raise an exception here, the launcher
will simply loop indefinitely.

To avoid this, we will add a new exception class to indicate a
permanent configuration error which was detected at runtime.  If
AWS says an instance type doesn't exist when we try to calculate
its quota, we mark it as permanently errored in the provider
configuration, then return empty quota information back to the
launcher.

This allows the launcher to accept the request, but then immediately
mark it as failed because the label isn't available.

The state of this error is stored on the provider manager, so the
typical corrective action of updating the configuration to correct
the label config means that a new provider will be spawned with an
empty error label list; the error state will be cleared and the
launcher will try again.

Finally, an extra exception handler is added to the main launcher
loop so that if any other unhandled errors slip through, the
request will be deferred and the launcher will continue processing
requests.

Change-Id: I9a5349203a337ab23159806762cb46c059fe4ac5