Always complete cache init on reconnect
There are some race conditions in the current treecache code that
could cause a cache to initialize empty or run without workers.

The following sequence can occur:

* ZK connection -> suspended
* Start cache init #1
* Cache init #1 starts workers
* ZK connection -> connected
* Set stop workers flag
* Workers stop
* Short spawn of cache init (call this cache init #2)
* Cache init #2 fails to obtain lock since init #1 is running
* Cache init #1 finishes and releases lock

This results in a cache with no worker threads running.

The problem is that we assume that if we reconnect to ZK multiple
times and a cache init is already running that we don't need to
run another one.  However, that depends greatly on the exact timing.

To fix this, we will always run the cache init each time we spawn
it, rather than assuming that an already running cache init is
sufficient.  If we do have a flapping connection, we may re-init
the cache more than necessary, but at least we should end up with
a working cache at the end.

Change-Id: I3ba56a5cddc516656bcce016ed048f4805ef3751