Use diskimage username in AWS and Azure drivers
The AWS and Azure drivers incorrectly required the user to supply
the username in the pool configuration when using diskimages.
The OpenStack and IBMVPC drivers correctly use the top-level
diskimage configuration to determine the username.

Correct this by deprecating the pool-level configuration in the
drivers that offer it, and default it to using the top-level
configuration.

Change-Id: I4e6b4d4268b32ab7b397a11dd0ccd08b18c09a86