Resolve statsd client once at startup
We currently create a new statsd client each time we replace a
provider manager.  This means that if we are unable to resolve
DNS at that time, the new provider may crash due to unhandled
exceptions.

To resolve this, let's adopt the same behavior we have in Zuul
which is to set up the statsd client once at startup and continue
to use the same client object for the life of the process.

This means that operators will still see errors at startup during
a misconfiguration, but any external changes after that will
not affect nodepool.

Change-Id: I65967c71e859fddbea15aee89f6ddae44344c87b