Add option to decline node request based on quota
Adds an option to control if a node request will be declined based on
available quota. This new option defaults to True to keep the existing
behavior; if set to False, a node request will be accepted even if there
is not enough quota to satisfy it, but may have to wait to be fulfilled.

Change-Id: Ic341cc4f34466d6b601e4eb57502aeab0e38ed82