Handle more quota errors during server creation
If during server creation the server ends up in error state due to
insufficient quota we would raise an exception and fail the node
request.

To fix this we will check the error message to detect quota related
errors and raise a `QuotaException` in those case to retry the request.

Error message was:

    ERROR nodepool.StateMachineNodeLauncher.provider: Error in creating
    the server. Compute service reports fault: Build of instance [...]
    aborted: VolumeSizeExceedsAvailableQuota: Requested volume or
    snapshot exceeds allowed gigabytes quota. Requested 480G, quota is
    102000G and 101880G has been consumed.

Change-Id: I5b8cf86ec9fff93920c82cb9e3abad8807cc1d7b