nodeutils: Log keyscan connection errors
When attempting to debug keyscan errors, it can be useful to know the
exact error (e.g. connection refused, host unreachable, or timeout) that
caused nodepool to try again. While not generally needed as it can flood
the logs with a lot of messages, log the reason at the debug log level
so it can be enabled if desired.

Change-Id: I6d52c94a03a3fbc9941b9211969bc88f8d09534c