Update patch set 3
Patch Set 3:

Thanks for the update.

What you describe is actually a bug in Zuul, so Zuul should be updated to handle deserialization failures.

In Nodepool, I'm pretty sure that there is a migration issue going from string to dict, so if you wanted to repropose this patch as, say, some kind of efficiency improvement, it would need to handle that.  But I don't think we should consider that until Zuul is fixed, so let's please fix it there first before doing any more work here.

FYI, the Node.user_data field is an opaque data field that is not meant to be understood by Nodepool.  The "user" in user_data refers to the requestor.  When Zuul accepts nodes, it stores some information in the user_data field for later use.

The metastatic nodepool driver is the *only* driver that reads/writes the Node.user_data field because it acts as a Nodepool requestor (requesting nodes from other Nodepool drivers).  Only when it is acting in the capacity of a nodepool user does it read and write the user_data field.  There are no other drivers to compare this behavior to, and therefore no right or wrong behavior.

Patch-set: 3
Attention: {"person_ident":"Gerrit User 31083 \u003c31083@4a232e18-c5a9-48ee-94c0-e04e7cca6543\u003e","operation":"ADD","reason":"Someone else replied on the change"}